Move `RemoteImageCommentsMapper` to separate file
