+++ EssentialFeed/EssentialFeed/Feed API/RemoteImageCommentsLoader.swift
@@ -88,27 +88,6 @@ public struct RemoteImageComment: Decodable {
 	}
 }
 
-public struct RemoteImageCommentsMapper {
-	private struct Root: Decodable {
-		let items: [RemoteImageComment]
-	}
-	
-	static func map(_ data: Data, from response: HTTPURLResponse) throws -> [RemoteImageComment] {
-		guard response.isOK,
-			  let root = try? iso8601Decoder().decode(Root.self, from: data) else {
-			throw RemoteImageCommentsLoader.Error.invalidData
-		}
-		
-		return root.items
-	}
-	
-	private static func iso8601Decoder() -> JSONDecoder {
-		let decoder = JSONDecoder()
-		decoder.dateDecodingStrategy = .iso8601
-		return decoder
-	}
-}
-
 private extension Array where Element == RemoteImageComment {
 	func toModels() -> [ImageComment] {
 		map { ImageComment(
